+++ domain/src/main/java/com/team_ia/domain/entity/GoogleLoginEntity.kt
@@ -2,8 +2,6 @@ package com.team_ia.domain.entity
 
 data class GoogleLoginEntity (
     val accessToken: String,
-    val expiresIn: Int,
-    val scope: String,
-    val tokenType: String,
-    val idToken: String
+    val refreshToken: String,
+    val expiredAt: Int
 )
