+++ domain/src/main/java/com/team_ia/domain/param/GoogleLoginParam.kt
@@ -1,13 +1,5 @@
 package com.team_ia.domain.param
 
 data class GoogleLoginParam(
-    val grantType: String,
-
-    val clientId: String,
-
-    val clientSecret: String,
-
-    val redirectUri: String,
-
     val code: String
 )
