+++ presentation/src/main/java/com/team_ia/idea_archive_android/ui/main/LoginActivity.kt
@@ -61,16 +61,11 @@ class LoginActivity : BaseActivity<ActivityLoginPageBinding>(R.layout.activity_l
             try {
                 val authCode = completedTask.getResult(ApiException::class.java)?.serverAuthCode
 
-                mainScope{
-                    authCode?.run{
-                        var accessToken = "userToken"
-                        loginViewModel.fetchGoogleAuthInfo(this).let {reult ->
-                            when (result) {
-                                is Result.Success<LoginGoogleResponse> -> {
+                mainScope {
+                    authCode?.run {
 
-                                }
-                            }
-                        }
+                    }
+
                     }
                 }
             }
